requirements.txt
amqp==2.5.1
Babel==2.6.0
billiard<4.0,>=3.6.0
celery>=4.2.1
certifi==2018.11.29
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.5
django-braces==1.13.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
drf-yasg==1.12.1
flex==6.13.2
flower==0.9.2
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
kombu<5.0,>=4.4.0
MarkupSafe==1.1.0
oauthlib==3.0.1
pika==0.13.0
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
qrcode==6.1
requests==2.21.0
requests-oauthlib==1.2.0
rfc3987==1.3.8
ruamel.yaml==0.15.87
six==1.12.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.3
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
vine>=1.3.0
django-restful-admin==1.0.3
twilio==6.27.1
channels==2.2.0
openpyxl==2.6.0
python-docx==0.8.10
docxtpl==0.11.4
xhtml2pdf==0.2.4
usps-api==0.5
